% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_pinguins.R
\name{filtrer_penguins}
\alias{filtrer_penguins}
\title{Filtrer le jeu de donnees penguins par espece}
\usage{
filtrer_penguins(data, espece_choisie)
}
\arguments{
\item{data}{Un data frame contenant les donnees penguins.}

\item{espece_choisie}{Une chaine de caracteres indiquant
l'espece a conserver.}
}
\value{
Un data frame filtre selon l'espece choisie.
}
\description{
Cette fonction conserve uniquement les lignes
correspondant a l'espece choisie.
}
