% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_pinguins.R
\name{graphique_longueur_bec_max}
\alias{graphique_longueur_bec_max}
\title{Representer la longueur maximale du bec par espece}
\usage{
graphique_longueur_bec_max(data)
}
\arguments{
\item{data}{Un data frame contenant les donnees penguins.}
}
\value{
Un graphique ggplot2 representant la longueur maximale
du bec par espece.
}
\description{
Cette fonction produit un graphique en barres
de la longueur maximale du bec des manchots pour chaque espece.
}
