% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_pinguins.R
\name{moyenne_poids_espece}
\alias{moyenne_poids_espece}
\title{Calculer le poids moyen par espece}
\usage{
moyenne_poids_espece(data)
}
\arguments{
\item{data}{Un data frame contenant les donnees penguins.}
}
\value{
Un data frame contenant une ligne par espece
et le poids moyen associe.
}
\description{
Cette fonction calcule le poids moyen des manchots
pour chaque espece presente dans le jeu de donnees.
}
