% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Jeu de donnees pinguins}
\format{
Un data frame avec 333 lignes et 8 variables :
\describe{
\item{espece}{Espece du manchot. Variable qualitative.}
\item{ile}{Ile d'observation. Variable qualitative.}
\item{longueur_bec}{Longueur du bec en millimetres. Variable quantitative.}
\item{profondeur_bec}{Profondeur du bec en millimetres. Variable quantitative.}
\item{longueur_nageoire}{Longueur de la nageoire en millimetres. Variable quantitative.}
\item{poids}{Poids du manchot en grammes. Variable quantitative.}
\item{sexe}{Sexe du manchot. Variable qualitative.}
\item{annee}{Annee d'observation. Variable quantitative discrete.}
}
}
\source{
Palmer Penguins, jeu de donnees prepare a partir du package
palmerpenguins.
}
\description{
Ce jeu de donnees contient des informations sur des manchots
observes sur plusieurs iles entre 2007 et 2009. Chaque ligne correspond
a un individu.
}
